Implemented Data Store Component
